Updated the README with central doc info.
Im not sure if we want to use the readme like this. eventually we are going to write up an actual readme on the game. I wasnt sure if people used or cared about the central doc, so I thought id put some of the progress info here. If we are using the doc, we dont have to merge this.